+++ Editor/Material Inspectors/OceanGUI.cs
@@ -1,0 +1,25 @@
+using UnityEditor;
+using UnityEditor.Rendering.HighDefinition;
+using UnityEngine;
+using UnityEngine.Rendering.HighDefinition;
+
+namespace GOcean
+{
+    public class OceanGUI : HDShaderGUI
+    {
+        MaterialUIBlockList uiBlocks = new MaterialUIBlockList
+        {
+            //new SurfaceOptionUIBlock(MaterialUIBlock.ExpandableBit.Base, features: SurfaceOptionUIBlock.Features.Lit),
+            //new TransparencyUIBlock(MaterialUIBlock.ExpandableBit.Transparency, features: TransparencyUIBlock.Features.All & ~TransparencyUIBlock.Features.Distortion),
+            //new EmissionUIBlock(MaterialUIBlock.ExpandableBit.Emissive),
+            //new AdvancedOptionsUIBlock(MaterialUIBlock.ExpandableBit.Advance, AdvancedOptionsUIBlock.Features.StandardLit)
+        };
+
+        protected override void OnMaterialGUI(MaterialEditor materialEditor, MaterialProperty[] props)
+        {
+            uiBlocks.OnGUI(materialEditor, props);
+        }
+
+        //public override void ValidateMaterial(Material material) => LitAPI.ValidateMaterial(material);
+    }
+}